requirements.txt
pytest==8.0.0
requests==2.32.3
cryptography==45.0.3
python-dotenv==1.0.0
allure-pytest==2.13.0